Moved common portions to core
