--- src/app/components/base/general.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { GeneralService } from './general.service';
-
-describe('GeneralService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: GeneralService = TestBed.get(GeneralService);
-    expect(service).toBeTruthy();
-  });
-});
