+++ src/app/core/core.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { GeneralService } from './services/general.service';
+import { WorkspaceService } from './services/workspace.service';
+
+
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule
+  ],
+  providers: [
+    GeneralService,
+    WorkspaceService
+  ]
+})
+export class CoreModule { }
