+++ src/app/core/guard/auth.guard.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, Router } from '@angular/router';
-import { AuthService } from './auth.service';
+import { AuthService } from '../../components/auth/auth.service';
  
 @Injectable({
   providedIn: 'root'
