+++ src/app/core/services/general.service.ts
