+++ src/app/core/services/workspace.service.ts
